pyproject.toml
[project]
name = "lightrag-spanner"
version = "0.1.0"
description = "Google Cloud Spanner storage backend for LightRAG"
requires-python = ">=3.10"
dependencies = [
    "lightrag-hku>=1.3.0",
    "google-cloud-spanner>=3.0.0",
    "numpy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lightrag_spanner"]
